plugins/OpenManus/requirements.txt
# 音频处理依赖
fish-audio-sdk>=1.0.0  # Fish Audio TTS（可选）
requests>=2.28.0       # MiniMax TTS使用

# 工具依赖
python-dateutil>=2.8.2  # 日期时间工具 
pandas>=1.5.3
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
akshare
bilix
firecrawl-py>=0.1.0    # Firecrawl 爬虫和搜索工具
aiohttp>=3.8.0         # 异步 HTTP 请求